requirements.txt
# Core web framework
streamlit>=1.28.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# AI and NLP libraries
google-genai>=0.5.0
spacy>=3.7.0
scikit-learn>=1.3.0
nltk>=3.8.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0

# Text processing and similarity
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
textstat>=0.7.0


# Web scraping and API calls
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data validation and parsing
pydantic>=2.4.0
dateparser>=1.1.0
phonenumbers>=8.13.0
email-validator>=2.1.0

# Visualization and reporting
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.13.0

# File handling
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development and testing
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

# Additional utilities
tqdm>=4.66.0
joblib>=1.3.0
